+++ app/Providers/AppServiceProvider.php
@@ -3,6 +3,9 @@
 namespace App\Providers;
 
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Support\Facades\Mail;
+use Symfony\Component\Mailer\Transport\Dsn;
+use GuzzleHttp\Client;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -13,6 +16,11 @@ class AppServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
-        //
+        Mail::extend('brevo', function () {
+            return new \App\Mail\Transport\BrevoTransport(
+                new Client(),
+                config('mail.mailers.brevo.key')
+            );
+        });
     }
 }