+++ storage/framework/views/c91cb7bace5c631d9f390e157585fb4d.php
@@ -1,0 +1,8 @@
+<div
+    <?php echo e($attributes->class(['fi-no-notification-body overflow-hidden break-words text-sm text-gray-500 dark:text-gray-400'])); ?>
+
+>
+    <?php echo e($slot); ?>
+
+</div>
+<?php /**PATH C:\Users\mate\OneDrive\Dokumenti\Cursor\Kornati\project-bolt-sb1-6vgshprr\project\vendor\filament\notifications\src\/../resources/views/components/body.blade.php ENDPATH**/ ?>